+++ src/main/java/roomescape/global/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.global.exception;
+package roomescape.global.exception;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.CONFLICT;
