+++ src/main/java/roomescape/global/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.global.exception;
+package roomescape.global.exception;
 
 import java.util.NoSuchElementException;
