+++ src/main/java/roomescape/reservation/infrastructure/db/ReservationThemeDbRepository.java
@@ -7,8 +7,8 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Repository;
 import roomescape.reservation.domain.entity.ReservationTheme;
 import roomescape.reservation.domain.repository.ReservationThemeRepository;
-import roomescape.reservation.global.exception.ResourceInUseException;
-import roomescape.reservation.global.exception.ResourceNotFoundException;
+import roomescape.global.exception.ResourceInUseException;
+import roomescape.global.exception.ResourceNotFoundException;
 import roomescape.reservation.infrastructure.db.dao.ReservationThemeDao;
 
 @Repository
