+++ src/test/java/roomescape/application/AdminReservationThemeServiceTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.reservation.application.AdminReservationThemeService;
-import roomescape.reservation.global.exception.BusinessRuleViolationException;
+import roomescape.global.exception.BusinessRuleViolationException;
 import roomescape.support.IntegrationTestSupport;
 
 class AdminReservationThemeServiceTest extends IntegrationTestSupport {
