+++ src/test/java/roomescape/application/AdminReservationTimeServiceTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.reservation.application.AdminReservationTimeService;
-import roomescape.reservation.global.exception.BusinessRuleViolationException;
+import roomescape.global.exception.BusinessRuleViolationException;
 import roomescape.support.IntegrationTestSupport;
 
 class AdminReservationTimeServiceTest extends IntegrationTestSupport {
