+++ src/test/java/roomescape/repository/ReservationThemeDbRepositoryTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.reservation.domain.entity.ReservationTheme;
-import roomescape.reservation.global.exception.ResourceNotFoundException;
+import roomescape.global.exception.ResourceNotFoundException;
 import roomescape.reservation.infrastructure.db.ReservationThemeDbRepository;
 import roomescape.reservation.infrastructure.db.dao.ReservationThemeH2Dao;
 import roomescape.support.JdbcTestSupport;
