+++ src/test/java/roomescape/repository/ReservationTimeDbRepositoryTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.reservation.domain.entity.ReservationTime;
-import roomescape.reservation.global.exception.ResourceNotFoundException;
+import roomescape.global.exception.ResourceNotFoundException;
 import roomescape.reservation.infrastructure.db.ReservationTimeDbRepository;
 import roomescape.reservation.infrastructure.db.dao.ReservationTimeH2Dao;
 import roomescape.support.JdbcTestSupport;
