+++ src/test/java/roomescape/service/UserReservationServiceTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.reservation.application.UserReservationService;
-import roomescape.reservation.global.exception.BusinessRuleViolationException;
+import roomescape.global.exception.BusinessRuleViolationException;
 import roomescape.reservation.domain.entity.Reservation;
 import roomescape.reservation.domain.repository.ReservationRepository;
 import roomescape.reservation.application.dto.request.CreateReservationServiceRequest;
